refactor(ios): inline the subscription bridge into KlaviyoBridge.swift
Folds createSubscription and its two consent mappers into KlaviyoBridge
itself and drops ios/KlaviyoSubscriptionBridge.swift. That file existed
only to keep KlaviyoBridge under SwiftLint's 400-line file_length
warning; per review, one file for the whole bridge reads better than an
extension split off for a lint metric.

KlaviyoBridge.swift is now 461 lines, so file_length warns. Accepted
deliberately: `yarn lint` runs swiftlint without --strict, so warnings do
not fail lint, and .swiftlint.yml already disables type_body_length on
the same reasoning.

The ObjC selector is unchanged, since the type is the same:
createSubscriptionWithSubscription: still resolves, so
KlaviyoReactNativeSdk.mm needs no edit.
